sidebar tooltip component seperated
