+++ react-practice copy/trycontextapi/src/App.css
@@ -38,238 +38,3 @@
   }
 }
 
-.modalmain {
-  background-color: beige;
-  display: flex;
-  justify-content: center;
-  align-items: center;
-  width: 100vw;
-  height: 100vh;
-  position: relative;
-
-}
-
-
-.cross {
-  position: absolute;
-  top: 0;
-  right: 1rem;
-  cursor: pointer;
-
-}
-
-.btn {
-  position: relative;
-  z-index: 1;
-  background-color: lightblue;
-  padding: 1rem;
-}
-
-
-.crossmodal {
-  background-color: lightblue;
-  display: flex;
-  justify-content: center;
-  align-items: center;
-  gap: 2rem;
-  border: 2px solid black;
-  width: 50vw;
-  height: 50vh;
-  flex-direction: column;
-  position: absolute;
-  text-align: center;
-  z-index: 3;
-  left: 50%;
-  top: 50%;
-  transform: translate(-50%, -50%);
-}
-
-.fade-in-image {
-  animation: fadeIn 0.5s forwards;
-}
-
-.fade-out-image {
-  animation: fadeOut 0.5s forwards;
-}
-
-@keyframes fadeIn {
-  0% {
-    opacity: 0;
-  }
-
-  50% {
-    opacity: 0.7;
-  }
-
-  100% {
-    opacity: 1;
-  }
-}
-
-@keyframes fadeOut {
-  0% {
-    opacity: 1;
-  }
-
-  50% {
-    opacity: 0.7;
-  }
-
-  100% {
-    opacity: 0;
-  }
-}
-
-
-
-.outdiv {
-  position: absolute;
-  top: 0;
-  left: 0;
-  right: 0;
-  bottom: 0;
-  background-color: rgba(0, 0, 0, 0.7);
-  display: flex;
-  justify-content: center;
-  align-items: center;
-  z-index: 2;
-
-}
-
-* {
-  padding: 0;
-  margin: 0;
-  box-sizing: border-box;
-}
-
-:root {
-  --color-blue: #4635b1;
-  --color-purple: #b771e5;
-  --color-green: #aeea94;
-  --color-yellow: #fffbca;
-}
-
-main {
-  display: grid;
-  grid-template-columns: repeat(3, 1fr);
-  grid-template-rows: repeat(3, 1fr);
-  background: var(--color-purple);
-  height: 100vh;
-  width: 100vw;
-  gap: 5px;
-}
-
-.card {
-  background-color: var(--color-yellow);
-  display: grid;
-  place-items: center;
-}
-
-button {
-  background-color: var(--color-green);
-  color: var(--color-blue);
-}
-
-span {
-  background-color: var(--color-purple);
-  color: var(--color-yellow);
-  padding: 4px 8px;
-  border-radius: 12px;
-  cursor: default;
-}
-
-.maindiv {
-  position: relative;
-  display: inline-block;
-}
-
-.showtool {
-  background-color: lightblue;
-  padding: 0.3rem;
-  position: absolute;
-  text-align: center;
-  border: 1px solid black;
-  color: black;
-  width: max-content;
-  min-width: 16px;
-  z-index: 1;
-}
-
-.top {
-  position: absolute;
-  top: -100%;
-  left: 50%;
-  transform: translate(-50%,-50%);
-}
-
-.bottom {
-  position:absolute;
-  bottom: -50%;
-  left: 50%;
-  transform: translate(-50%,100%);
-}
-
-.left {
-  position: absolute;
-  left: -40%;
-  bottom: -20%;
-  transform: translate(-50%,0%);
-}
-
-.right {
-  position: absolute;
-  right: -40%;
-  bottom: -20%;
-  transform: translate(50%,0%);
-}
-
-.sidebar {
-  position: absolute;
-  top: 0;
-  right: -250px;
-  width: 300px;
-  height: 100%;
-  background-color: lightblue;
-  padding: 20px;
-  display: block;
-  z-index: 2;
-  color:black;
-  text-align: center;
-  justify-content: center;
-  align-items: center;
-  display: flex;
-  animation: slideIn 0.5s ease-out forwards;
-}
-
-.slideOut{
-  animation: slideOut 0.5s ease-out forwards;
-}
-
-.Sidebar{
-  position: absolute;
-  width: 100vw;
-  height: 100vh;
-  background-color: rgba(0,0,0,0.6);
-  z-index: 1;
-}
-
-@keyframes slideIn {
-  0% {
-    right: -250px; 
-  }
-  100% {
-    right: 0; 
-  }
-}
-
-
-@keyframes slideOut {
-  0% {
-    right: 0;
-  }
-  100% {
-    right: -350px;
-  }
-}
-
-
