+++ react-practice copy/trycontextapi/src/ToolTips/Tooltip.css
@@ -1,0 +1,182 @@
+ 
+  * {
+    padding: 0;
+    margin: 0;
+    box-sizing: border-box;
+  }
+  
+  :root {
+    --color-blue: #4635b1;
+    --color-purple: #b771e5;
+    --color-green: #aeea94;
+    --color-yellow: #fffbca;
+  }
+  
+  main{
+    display: grid;
+    grid-template-columns: repeat(3, 1fr);
+    grid-template-rows: repeat(3, 1fr);
+    background: var(--color-purple);
+    height: 100vh;
+    width: 100vw;
+    gap: 5px;
+  }
+  
+  .card {
+    background-color: var(--color-yellow);
+    display: grid;
+    place-items: center;
+  }
+  
+  button {
+    background-color: var(--color-green);
+    color: var(--color-blue);
+  }
+  
+  span {
+    background-color: var(--color-purple);
+    color: var(--color-yellow);
+    padding: 4px 8px;
+    border-radius: 12px;
+    cursor: default;
+  }
+  
+  .maindiv {
+    position: relative;
+    display: inline-block;
+  }
+  
+  .showtool {
+    background-color: lightblue;
+    padding: 0.3rem;
+    position: absolute;
+    text-align: center;
+    border: 1px solid black;
+    color: black;
+    width: max-content;
+    min-width: 16px;
+    z-index: 1;
+  }
+  
+  .top {
+    position: absolute;
+    top: -100%;
+    left: 50%;
+    transform: translate(-50%,-50%);
+  }
+  
+  .bottom {
+    position:absolute;
+    bottom: -50%;
+    left: 50%;
+    transform: translate(-50%,100%);
+  }
+  
+  .left {
+    position: absolute;
+    left: -40%;
+    bottom: -20%;
+    transform: translate(-50%,0%);
+  }
+  
+  .right {
+    position: absolute;
+    right: -40%;
+    bottom: -20%;
+    transform: translate(50%,0%);
+  }
+  
+  .sidebar {
+    position: absolute;
+    top: 0;
+    right: -250px;
+    width: 300px;
+    height: 100%;
+    background-color: lightblue;
+    padding: 20px;
+    display: block;
+    z-index: 2;
+    color:black;
+    text-align: center;
+    justify-content: center;
+    align-items: center;
+    display: flex;
+    animation: slideIn 0.5s ease-out forwards;
+  }
+  
+  .slideOut{
+    animation: slideOut 0.5s ease-out forwards;
+  }
+  
+  .Sidebar{
+    position: absolute;
+    width: 100vw;
+    height: 100vh;
+    background-color: rgba(0,0,0,0.6);
+    z-index: 1;
+  }
+  
+  @keyframes slideIn {
+    0% {
+      right: -250px; 
+    }
+    100% {
+      right: 0; 
+    }
+  }
+  
+  
+  @keyframes slideOut {
+    0% {
+      right: 0;
+    }
+    100% {
+      right: -350px;
+    }
+  }
+
+  :root {
+    font-family: Inter, system-ui, Avenir, Helvetica, Arial, sans-serif;
+    line-height: 1.5;
+    font-weight: 400;
+  
+    color-scheme: light dark;
+    color: rgba(255, 255, 255, 0.87);
+    background-color: #242424;
+  
+    font-synthesis: none;
+    text-rendering: optimizeLegibility;
+    -webkit-font-smoothing: antialiased;
+    -moz-osx-font-smoothing: grayscale;
+  }
+  
+  body {
+    margin: 0;
+    display: flex;
+    place-items: center;
+    min-width: 320px;
+    min-height: 100vh;
+  }
+  
+  h1 {
+    font-size: 3.2em;
+    line-height: 1.1;
+  }
+  
+  button {
+    border-radius: 8px;
+    border: 1px solid transparent;
+    padding: 0.6em 1.2em;
+    font-size: 1em;
+    font-weight: 500;
+    font-family: inherit;
+    cursor: pointer;
+    transition: border-color 0.25s;
+  }
+  button:focus,
+  button:focus-visible {
+    background-color: lightgreen;
+  }
+  
+  
+
